Make RANGE() return a float, else the range might not fit into the data type.
git-svn-id: svn://svn.berlios.de/gale/trunk@1397 2d798ef8-6e06-0410-941a-9092aa770104

+++ gale2/include/gale/math/colormodel.h
@@ -48,9 +48,9 @@ class ColorModel
     template<unsigned int N,typename T>
     void init(Color<N,T> const& color) {
         // Map RGB to range [0,1].
-        float r=clamp(static_cast<float>(color.getR()-color.MIN_VALUE())/color.RANGE(),0.0f,1.0f);
-        float g=clamp(static_cast<float>(color.getG()-color.MIN_VALUE())/color.RANGE(),0.0f,1.0f);
-        float b=clamp(static_cast<float>(color.getB()-color.MIN_VALUE())/color.RANGE(),0.0f,1.0f);
+        float r=clamp((color.getR()-color.MIN_VALUE())/color.RANGE(),0.0f,1.0f);
+        float g=clamp((color.getG()-color.MIN_VALUE())/color.RANGE(),0.0f,1.0f);
+        float b=clamp((color.getB()-color.MIN_VALUE())/color.RANGE(),0.0f,1.0f);
 
         // Convert to the color model.
         fromRGB(r,g,b);
